feat: give tournament edit for local and published tournaments a separate url
