+++ src/app/modules/tournament-management/components-utility/tournament-card/tournament-card.component.html
@@ -1,4 +1,4 @@
-<div class="tournament" (click)="editTournament(tournament, $event)">
+<div class="tournament">
 	<div class="header">
 		<div class="information">
 			<div class="id" matTooltip="Tournament import id" *ngIf="tournament.publishId != null">
