+++ src/app/modules/tournament-management/tournament-management.module.ts
@@ -19,7 +19,7 @@ import { TournamentCardComponent } from './components-utility/tournament-card/to
 import { TournamentCreateComponent } from './pages/tournament-create/tournament-create.component';
 import { TournamentWybinComponent } from './components-utility/tournament/tournament-wybin/tournament-wybin.component';
 import { TournamentConditionalMessageComponent } from './components-utility/tournament/tournament-beatmap-result/tournament-conditional-message.component';
-import { TournamentEditComponent } from './components/tournament-manage/tournament-edit/tournament-edit.component';
+import { TournamentEditComponent } from './pages/tournament-edit/tournament-edit.component';
 import { LocalTournamentsComponent } from './pages/local-tournaments/local-tournaments.component';
 import { PublishedTournamentsComponent } from './pages/published-tournaments/published-tournaments.component';
